[FIX] sale_commission_salesman: partner fetch for agents creation on move lines
Fetch the partner by move line record instead of fetching all partners
for all invoice users related to the moves. The previous method could
fetch more than one partner, including those not related to the move,
when multiple lines for different moves are being processed, causing an
error when trying to obtain the values of the `agent` and
`salesman_as_agent` fields.